Move LLM catalogue and config into praisonai-code (C7).
Model listing and AutoGen config_list building no longer import the wrapper;
shims preserve praisonai.llm.* module identity for backward compatibility.
