+++ src/praisonai-code/praisonai_code/llm/__init__.py
@@ -9,4 +9,7 @@ __all__ = [
     "inject_credentials_into_env",
     "is_configured",
     "resolve_llm_endpoint_with_credentials",
+    "ModelCatalogue",
+    "ModelInfo",
+    "build_config_list",
 ]
